+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.1] - 2020-09-10
+## [1.2.0] - 2020-09-10
+### Added
+- Added `Palette` property to `IPixelTransform`.
+
+### Fixed
+- `NopPixelTransform` correctly made sealed.
+
+## [1.1.0] - 2020-09-10
 ### Added
 - Added `PaletteSize` property to `IPixelTransform`.